Refresh bug, refresh indicator and none info indicator

+++ webapp/src/components/windowComponents/friends/FriendsComponent.tsx
@@ -1,4 +1,4 @@
-import { Box, Divider, Tooltip } from '@mui/material'
+import { Box, CircularProgress, Divider, Tooltip } from '@mui/material'
 import React, { createRef, useEffect, useState } from 'react'
 import { styled } from '@mui/material/styles';
 import List from '@mui/material/List';
@@ -23,14 +23,19 @@ const VerticalDivider = styled(Divider)({
     padding: '0em 0.4em 0em'
 })
 
-const ScrollBox = styled(Box)({
-    maxHeight: '60vh',
-    overflow: 'auto',
-    scrollbarColor: 'black white'
+const BoxCircularProgress = styled(Box)({
+    display: 'flex',
+    justifyContent: 'center',
+    alignContent: 'center'
+})
+
+const InfoBox = styled(Box)({
+    color: '#1f4a21',
+    textAlign: 'center'
 })
 
 
-export const FriendsComponent = (props: { friends: Promise<Friend[]>, daddy: any, refresh: any }) => {
+export const FriendsComponent = (props: { friends: Promise<Friend[]>, daddy: any, refresh: any, stopLoading: any }) => {
 
     const [url, setUrl] = useState("../testUser.jfif");
 
@@ -62,10 +67,29 @@ export const FriendsComponent = (props: { friends: Promise<Friend[]>, daddy: any
     }
 
     const deleteFriend = (friend: User) => {
-        deleteFriendApi(friend).then(() => {
-            props.refresh()
-            temporalSuccessMessage("Tú amigo <em>" + friend.username + "</em> ha sido eliminado correctamente.");
-        });
+        showQuestion(friend)
+    }
+
+    const showQuestion = (friend: User) => {
+        Swal.fire({
+            title: "Eliminar amigo",
+            html: `<p>¿Esta seguro de que quieres eliminar al amigo <em><b> ${friend.username}</b></em>?</p>`,
+            icon: 'question',
+            showCancelButton: true,
+            confirmButtonColor: '#81c784',
+            cancelButtonColor: 'grey',
+            confirmButtonText: 'Eliminar',
+            cancelButtonText: 'Volver'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteFriendApi(friend).then(() => {
+                    props.refresh()
+                    temporalSuccessMessage("Tú amigo <em><b>" + friend.username + "</b></em> ha sido eliminado correctamente.");
+                });
+            } else {
+                Swal.close();
+            }
+        })
     }
 
     const showFriendProfile = async (user: User) => {
@@ -94,48 +118,57 @@ export const FriendsComponent = (props: { friends: Promise<Friend[]>, daddy: any
         try {
             render(
                 <>
-                    {frds.map((friend, i) => {
-                        return (
-                            <React.Fragment key={i}>
-                                <ListItemButton>
-                                    <ListItemIcon>
-                                        <Tooltip title="See friend profile">
-                                            <PersonIcon onClick={() => showFriendProfile(friend.user)} />
-                                        </Tooltip>
-                                    </ListItemIcon>
-                                    <ListItemText primary={friend.user.username} />
-                                    {isOpen(i) ?
-                                        <ExpandLess onClick={() => handleClick(i)} />
-                                        :
-                                        <ExpandMore onClick={() => handleClick(i)} />
-                                    }
-                                    <VerticalDivider orientation='vertical' flexItem />
-                                    <Tooltip title="Delete friend" sx={{ ml: "0.5em" }}>
-                                        <CloseIcon onClick={() => deleteFriend(friend.user)} htmlColor="red" />
-                                    </Tooltip>
-                                </ListItemButton>
-                                <Collapse in={isOpen(i)} timeout="auto" unmountOnExit>
-                                    <List component="div" disablePadding>
-                                        {friend.groups.map((group, j) => {
-                                            return (
-                                                <React.Fragment key={i + "-" + j}>
-                                                    <Tooltip title="Show on map">
-                                                        <ListItemButton sx={{ pl: 4 }} onClick={() => showGroup(group, friend.user)}>
-                                                            <ListItemIcon>
-                                                                <MapIcon />
-                                                            </ListItemIcon>
-                                                            <ListItemText primary={group.name} />
-                                                        </ListItemButton>
-                                                    </Tooltip>
-                                                </React.Fragment>
-                                            )
-                                        })}
-                                    </List>
-                                </Collapse>
-                            </React.Fragment>)
-                    })}
-                </ >, props.daddy.current)
+                    {frds.length > 0 ?
+                        <React.Fragment key="friend-render">
+                            {frds.map((friend, i) => {
+                                return (
+                                    <React.Fragment key={i}>
+                                        <ListItemButton>
+                                            <ListItemIcon>
+                                                <Tooltip title="See friend profile">
+                                                    <PersonIcon onClick={() => showFriendProfile(friend.user)} />
+                                                </Tooltip>
+                                            </ListItemIcon>
+                                            <ListItemText primary={friend.user.username} />
+                                            {isOpen(i) ?
+                                                <ExpandLess onClick={() => handleClick(i)} />
+                                                :
+                                                <ExpandMore onClick={() => handleClick(i)} />
+                                            }
+                                            <VerticalDivider orientation='vertical' flexItem />
+                                            <Tooltip title="Delete friend" sx={{ ml: "0.5em" }}>
+                                                <CloseIcon onClick={() => deleteFriend(friend.user)} htmlColor="red" />
+                                            </Tooltip>
+                                        </ListItemButton>
+                                        <Collapse in={isOpen(i)} timeout="auto" unmountOnExit>
+                                            <List component="div" disablePadding>
+                                                {friend.groups.map((group, j) => {
+                                                    return (
+                                                        <React.Fragment key={i + "-" + j}>
+                                                            <Tooltip title="Show on map">
+                                                                <ListItemButton sx={{ pl: 4 }} onClick={() => showGroup(group, friend.user)}>
+                                                                    <ListItemIcon>
+                                                                        <MapIcon />
+                                                                    </ListItemIcon>
+                                                                    <ListItemText primary={group.name} />
+                                                                </ListItemButton>
+                                                            </Tooltip>
+                                                        </React.Fragment>
+                                                    )
+                                                })}
+                                            </List>
+                                        </Collapse>
+                                    </React.Fragment>)
+                            })}
+                        </React.Fragment>
+                        :
+                        <InfoBox>
+                            <p><b>No tienes amigos añadidos.</b></p><p>¡Corre a preguntarles su nombre de usuario!</p>
+                        </InfoBox>
+                    }
+                </>, props.daddy.current)
+            props.stopLoading()
         } catch (e: any) { }
     })
-    return <></>
+    return (<></>)
 }
